Statistics
Added statitistic for persons/invoices

+++ invoice-server-starter/Invoices.Api/Models/InvoiceFilterDto.cs
@@ -2,8 +2,8 @@
 {
     public class InvoiceFilterDto
     {
-        public uint? sellerId {  get; set; }
-        public uint? buyerId { get; set; }
+        public ulong? sellerId {  get; set; }
+        public ulong? buyerId { get; set; }
         public string? product { get; set; }
         public decimal? minPrice { get; set; }
         public decimal? maxPrice { get; set; }
